WIP - Still working on the theming and accessibility feature. Fixed issue #129 (all windows are now movable).
