+++ Common/Common.csproj
@@ -33,8 +33,8 @@
     <PackageReference Include="System.Diagnostics.Debug" Version="4.3.0" />
     <PackageReference Include="System.Runtime.Extensions" Version="4.3.1" />
     <PackageReference Include="System.Net.NameResolution" Version="4.3.0" />
-    <PackageReference Include="Wokhan.Core" Version="0.9.2-alpha" />
-    <PackageReference Include="Wokhan.UI" Version="0.9.2-alpha" />
+    <PackageReference Include="Wokhan.Core" Version="0.9.4-beta" />
+    <PackageReference Include="Wokhan.UI" Version="0.9.6-beta" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Interop.NetFwTypeLib">
