+++ Common/UI/Theme.xaml
@@ -1,6 +1,9 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:ext="clr-namespace:Wokhan.UI.Xaml.Extensibility;assembly=Wokhan.UI">
+
+    <SolidColorBrush x:Key="AccentColorBrush" Color="{DynamicResource SystemParameters.WindowGlassColor}" />
+
     <Style TargetType="CheckBox">
         <Setter Property="OverridesDefaultStyle" Value="True"/>
         <Setter Property="Background" Value="{DynamicResource ButtonBackground}"/>
@@ -76,23 +79,17 @@
     </Style>
 
     <Style TargetType="ComboBox">
-        <Setter Property="ext:CustomAdorner.PreserveTemplate" Value="true" />
+        <Setter Property="ext:CustomAdorner.PreserveTemplate" Value="True" />
         <Setter Property="IsEditable" Value="False" />
         <!--<Setter Property="Background" Value="{DynamicResource ButtonBackground}" />-->
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="ComboBox">
                     <Grid>
-                        <ToggleButton ext:CustomAdorner.HeaderImage="{Binding Path=(ext:CustomAdorner.HeaderImage),RelativeSource={x:Static RelativeSource.TemplatedParent}}" ext:CustomAdorner.Header="{Binding Path=(ext:CustomAdorner.Header),RelativeSource={x:Static RelativeSource.TemplatedParent}}" Content="{Binding Text,RelativeSource={x:Static RelativeSource.TemplatedParent}}" ext:CustomAdorner.Image="{Binding Path=(ext:CustomAdorner.Image),RelativeSource={x:Static RelativeSource.TemplatedParent}}" IsChecked="{Binding IsDropDownOpen, Mode=TwoWay, RelativeSource={x:Static RelativeSource.TemplatedParent}}">
-                            <ToggleButton.Style>
-                                <Style TargetType="ToggleButton" BasedOn="{StaticResource {x:Type ToggleButton}}">
-                                    <Setter Property="Background" Value="{Binding Background,RelativeSource={x:Static RelativeSource.TemplatedParent}}" />
-                                </Style>
-                            </ToggleButton.Style>
-                        </ToggleButton>
+                        <ToggleButton ext:CustomAdorner.HeaderImage="{Binding Path=(ext:CustomAdorner.HeaderImage),RelativeSource={x:Static RelativeSource.TemplatedParent}}" ext:CustomAdorner.Header="{Binding Path=(ext:CustomAdorner.Header),RelativeSource={x:Static RelativeSource.TemplatedParent}}" Content="{Binding Text,RelativeSource={x:Static RelativeSource.TemplatedParent}}" ext:CustomAdorner.Image="{Binding Path=(ext:CustomAdorner.Image),RelativeSource={x:Static RelativeSource.TemplatedParent}}" IsChecked="{Binding IsDropDownOpen, Mode=TwoWay, RelativeSource={x:Static RelativeSource.TemplatedParent}}" />
                         <Popup x:Name="PART_Popup" IsOpen="{TemplateBinding IsDropDownOpen}" StaysOpen="False" AllowsTransparency="True" Focusable="False" PopupAnimation="Slide" Placement="Bottom">
-                            <Grid Background="{TemplateBinding Background}" MinWidth="{TemplateBinding FrameworkElement.ActualWidth}" MaxHeight="{TemplateBinding MaxDropDownHeight}" Margin="4,0,4,4">
-                                <Border Background="{TemplateBinding Background}" CornerRadius="4">
+                            <Grid MinWidth="{TemplateBinding FrameworkElement.ActualWidth}" MaxHeight="{TemplateBinding MaxDropDownHeight}" Margin="4,0,4,4">
+                                <Border Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" CornerRadius="4">
                                     <Border.Effect>
                                         <DropShadowEffect Color="LightGray" ShadowDepth="2" />
                                     </Border.Effect>
@@ -110,4 +107,19 @@
         </Setter>
     </Style>
 
+    <SolidColorBrush x:Key="WindowBrushSaved" Color="{x:Static SystemColors.WindowColor}" />
+    <SolidColorBrush x:Key="WindowTextBrushSaved" Color="{x:Static SystemColors.WindowTextColor}" />
+    <SolidColorBrush x:Key="ActiveCaptionBrushSaved" Color="{x:Static SystemColors.ActiveCaptionColor}" />
+    <SolidColorBrush x:Key="ScrollBarBrushSaved" Color="{x:Static SystemColors.ScrollBarColor}" />
+    <SolidColorBrush x:Key="MenuBarBrushSaved" Color="{x:Static SystemColors.MenuBarColor}" />
+    <SolidColorBrush x:Key="ControlTextBrushSaved" Color="{x:Static SystemColors.ControlTextColor}" />
+    <SolidColorBrush x:Key="ControlBrushSaved" Color="{x:Static SystemColors.ControlColor}" />
+    <SolidColorBrush x:Key="InfoBrushSaved" Color="{x:Static SystemColors.InfoColor}" />
+
+
+    <Style x:Key="SystemThemeStyle" TargetType="ContentControl">
+        <Style.Resources>
+            
+        </Style.Resources>
+    </Style>
 </ResourceDictionary>
