+++ Common/UI/Themes/Dark.xaml
@@ -1,34 +1,40 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:system="clr-namespace:System;assembly=System.Runtime">
+
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="../Theme.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+
     <SolidColorBrush x:Key="AccentColorBrush" Color="Coral" />
     <SolidColorBrush x:Key="DataGridCellSelectedForeground" Color="Black" />
     <SolidColorBrush x:Key="DataGridRowSelectedBackground" Color="SlateGray" />
     <SolidColorBrush x:Key="DataGridRowAltBackground" Color="#101010" />
-    <SolidColorBrush x:Key="HelpButtonOverBackground" Color="WhiteSmoke" />
 
     <SolidColorBrush x:Key="TabItemOverBackground" Color="#202020" />
 
     <SolidColorBrush x:Key="ButtonBackground" Color="#303030" />
-    <SolidColorBrush x:Key="ButtonOverBackground" Color="#202020" />
+    <SolidColorBrush x:Key="ButtonOverBackground" Color="#505050" />
 
-    <SolidColorBrush x:Key="StatusScreenPrivateBackgroundBrush" Color="DarkMagenta" />
-    <SolidColorBrush x:Key="StatusScreenPublicBackgroundBrush" Color="DarkGreen" />
-    <SolidColorBrush x:Key="StatusScreenDomainBackgroundBrush" Color="DarkBlue" />
-    <SolidColorBrush x:Key="StatusScreenAllBackgroundBrush" Color="Gray" />
+    <SolidColorBrush x:Key="StatusScreenPrivateBackgroundBrush" Color="DarkMagenta" Opacity="0.2" />
+    <SolidColorBrush x:Key="StatusScreenPublicBackgroundBrush" Color="DarkGreen" Opacity="0.2" />
+    <SolidColorBrush x:Key="StatusScreenDomainBackgroundBrush" Color="DarkBlue" Opacity="0.2" />
+    <SolidColorBrush x:Key="StatusScreenAllBackgroundBrush" Color="Gray" Opacity="0.2" />
 
-    
+    <system:Double x:Key="{x:Static SystemParameters.CaptionHeightKey}">38</system:Double>
+
     <SolidColorBrush x:Key="{x:Static SystemColors.WindowBrushKey}" Color="Black" />
     <SolidColorBrush x:Key="{x:Static SystemColors.WindowTextBrushKey}" Color="White" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.ActiveCaptionBrushKey}" Color="#202020" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ScrollBarBrushKey}" Color="#202020" />
     <SolidColorBrush x:Key="{x:Static SystemColors.MenuBarBrushKey}" Color="#101010" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ControlTextBrushKey}" Color="White" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ControlBrushKey}" Color="#202020" />
     <SolidColorBrush x:Key="{x:Static SystemColors.InfoBrushKey}" Color="Gray" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.InfoTextBrushKey}" Color="Black" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.HighlightBrushKey}" Color="Gray" />
     <!--<SolidColorBrush x:Key="{x:Static SystemColors.ButtonHighlightBrushKey}" Color="#202020" />-->
 
     <SolidColorBrush x:Key="{x:Static SystemColors.ControlLightBrushKey}" Color="Black" />
 
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="../Theme.xaml" />
-    </ResourceDictionary.MergedDictionaries>
 </ResourceDictionary>
