+++ Common/UI/Themes/Light.xaml
@@ -1,10 +1,16 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
+                    xmlns:system="clr-namespace:System;assembly=mscorlib">
+
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="../Theme.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+
     <SolidColorBrush x:Key="AccentColorBrush" Color="Coral" />
     <SolidColorBrush x:Key="DataGridCellSelectedForeground" Color="Black" />
     <SolidColorBrush x:Key="DataGridRowSelectedBackground" Color="LightBlue" />
-    <SolidColorBrush x:Key="HelpButtonOverBackground" Color="WhiteSmoke" />
-    
+    <SolidColorBrush x:Key="DataGridRowAltBackground" Color="LightGray" />
+
     <SolidColorBrush x:Key="TabItemOverBackground" Color="WhiteSmoke" />
     
     <SolidColorBrush x:Key="ButtonBackground" Color="WhiteSmoke" />
@@ -15,16 +21,20 @@
     <SolidColorBrush x:Key="StatusScreenDomainBackgroundBrush" Color="#FFC1F4FF" />
     <SolidColorBrush x:Key="StatusScreenAllBackgroundBrush" Color="LightGray" />
 
+    <system:Double x:Key="{x:Static SystemParameters.CaptionHeightKey}">38</system:Double>
 
     <SolidColorBrush x:Key="{x:Static SystemColors.WindowBrushKey}" Color="WhiteSmoke" />
     <SolidColorBrush x:Key="{x:Static SystemColors.WindowTextBrushKey}" Color="Black" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.ActiveCaptionBrushKey}" Color="LightGray" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ScrollBarBrushKey}" Color="WhiteSmoke" />
     <SolidColorBrush x:Key="{x:Static SystemColors.MenuBarBrushKey}" Color="LightGray" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ControlTextBrushKey}" Color="Black" />
     <SolidColorBrush x:Key="{x:Static SystemColors.ControlBrushKey}" Color="White" />
     <SolidColorBrush x:Key="{x:Static SystemColors.InfoBrushKey}" Color="LightYellow" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.InfoTextBrushKey}" Color="Black" />
+    <SolidColorBrush x:Key="{x:Static SystemColors.HighlightBrushKey}" Color="Gray" />
+    <!--<SolidColorBrush x:Key="{x:Static SystemColors.ButtonHighlightBrushKey}" Color="#202020" />-->
 
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="../Theme.xaml" />
-    </ResourceDictionary.MergedDictionaries>
+    <SolidColorBrush x:Key="{x:Static SystemColors.ControlLightBrushKey}" Color="Black" />
+    
 </ResourceDictionary>
