+++ Common/UI/Themes/System.xaml
@@ -1,21 +1,26 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:system="clr-namespace:System;assembly=System.Runtime">
+    
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="../Theme.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+
     <SolidColorBrush x:Key="AccentColorBrush" Color="Coral" />
+
     <SolidColorBrush x:Key="DataGridCellSelectedForeground" Color="Black" />
     <SolidColorBrush x:Key="DataGridRowSelectedBackground" Color="LightBlue" />
-    <SolidColorBrush x:Key="HelpButtonOverBackground" Color="WhiteSmoke" />
 
     <SolidColorBrush x:Key="TabItemOverBackground" Color="{x:Static SystemColors.ControlColor}" />
 
-    <SolidColorBrush x:Key="ButtonBackground" Color="{x:Static SystemColors.ControlColor}" />
+    <SolidColorBrush x:Key="ButtonBackground" Color="{x:Static SystemColors.ControlDarkColor}" />
     <SolidColorBrush x:Key="ButtonOverBackground" Color="{x:Static SystemColors.ControlLightColor}" />
+    
+    <system:Double x:Key="{x:Static SystemParameters.CaptionHeightKey}">38</system:Double>
 
     <SolidColorBrush x:Key="StatusScreenPrivateBackgroundBrush" Color="{x:Static SystemColors.ControlColor}" />
     <SolidColorBrush x:Key="StatusScreenPublicBackgroundBrush" Color="{x:Static SystemColors.ControlColor}" />
     <SolidColorBrush x:Key="StatusScreenDomainBackgroundBrush" Color="{x:Static SystemColors.ControlColor}" />
     <SolidColorBrush x:Key="StatusScreenAllBackgroundBrush" Color="{x:Static SystemColors.ControlColor}" />
-
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="../Theme.xaml" />
-    </ResourceDictionary.MergedDictionaries>
+   
 </ResourceDictionary>
