+++ Console/UI/Windows/About.xaml
@@ -1,0 +1,78 @@
+﻿<Window x:Class="Wokhan.WindowsFirewallNotifier.Console.UI.Windows.About"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:ext="clr-namespace:Wokhan.UI.Xaml.Extensibility;assembly=Wokhan.UI"
+      xmlns:uicontrols="clr-namespace:Wokhan.UI.Xaml.Controls;assembly=Wokhan.UI"
+      DataContext="{Binding RelativeSource={x:Static RelativeSource.Self}}"
+      mc:Ignorable="d" 
+      Height="400" Width="650"
+      ShowInTaskbar="False"
+      ResizeMode="NoResize"
+      WindowStartupLocation="CenterOwner"
+      Title="About" Margin="10" ScrollViewer.CanContentScroll="True" ScrollViewer.VerticalScrollBarVisibility="Auto">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition />
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+        <uicontrols:CustomWindowChrome Height="38" DockPanel.Dock="Top" Background="{DynamicResource {x:Static SystemColors.ActiveCaptionBrushKey}}" />
+        <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto">
+            <StackPanel Margin="20">
+                <Grid Margin="0,0,0,20">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="Auto" />
+                        <ColumnDefinition />
+                    </Grid.ColumnDefinitions>
+                    <Border HorizontalAlignment="Left" Margin="10,0" Padding="2" VerticalAlignment="Center" Background="White" BorderBrush="Gray" BorderThickness="1">
+                        <Image Source="/WFN;component/Resources/Images/WFN.ico" Height="60" HorizontalAlignment="Left" />
+                    </Border>
+                    <StackPanel Grid.Column="1" VerticalAlignment="Center">
+                        <TextBlock FontSize="13" VerticalAlignment="Center" FontWeight="Bold">
+                            <Run Text="{Binding ProductName,Mode=OneTime,FallbackValue='WFN'}" />
+                            <Run Text="{Binding Version,Mode=OneTime,FallbackValue='X.X.X.X'}" />
+                        </TextBlock>
+                        <TextBlock Margin="0,5" TextWrapping="Wrap">
+                            <Run Text="Developed by" />
+                            <Hyperlink RequestNavigate="OpenExernalLink" NavigateUri="https://github.com/wokhansoft/WFN/graphs/contributors">the WFN contributors</Hyperlink>
+                            <LineBreak />
+                            <Run Text="Created by Jérôme Saliba (@wokhan), maintained for years by Daniel Pharros (@atlashackert), improved by Harry Wissiak (@harrwiss)." />
+                            <LineBreak />
+                            <Run FontWeight="Bold" Text="In memory of Harry Wissiak (@harrwiss), who passed away on June 6th 2020." />
+                        </TextBlock>
+                    </StackPanel>
+                </Grid>
+                <TextBlock Text="Description" Style="{StaticResource SectionHeader}" />
+                <TextBlock Margin="10" TextWrapping="Wrap" TextAlignment="Justify" Text="WFN offers network related tools and optionally extends the Windows embedded firewall behavior by displaying notifications to enable users to control outgoing connections (by default, Windows allows all of them, meaning that any program on your computer can send information to the Internet without your consent)." />
+                <TextBlock Text="License" Style="{StaticResource SectionHeader}" />
+                <TextBlock Margin="10" TextWrapping="Wrap" TextAlignment="Justify">This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+<LineBreak />
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ GNU General Public License for more details.
+<LineBreak/>
+ You should have received a copy of the GNU General Public License
+ along with this program. If not, see http://www.gnu.org/licenses/.</TextBlock>
+                <TextBlock Text="Donations" Style="{StaticResource SectionHeader}" />
+                <TextBlock Margin="10"  TextAlignment="Justify" TextWrapping="Wrap">
+            <Run Text="WFN is free and will always be. However, if you enjoy using it, you can consider donating using the Sponsor button on" />
+            <Hyperlink RequestNavigate="OpenExernalLink" NavigateUri="https://github.com/wokhan/wfn">GitHub</Hyperlink><Run Text=". Thanks!" />
+                </TextBlock>
+                <TextBlock Text="External libraries and components" Style="{StaticResource SectionHeader}" />
+                <StackPanel Margin="10">
+                    <TextBlock TextAlignment="Justify" TextWrapping="Wrap" Text="WFN uses XAML icons from http://modernuiicons.com and http://materialdesignicons.com." />
+                    <TextBlock Margin="0,10" TextAlignment="Justify" TextWrapping="Wrap" Text="This product includes GeoLite2 data created by MaxMind, available from http://www.maxmind.com." />
+                </StackPanel>
+            </StackPanel>
+        </ScrollViewer>
+        <ContentControl Grid.Row="2" Height="50" DockPanel.Dock="Bottom" Style="{StaticResource ToolBarPanel}">
+            <Button Click="Close" ext:CustomAdorner.Image="{StaticResource Image_Refresh}" Content="Close" VerticalAlignment="Top" HorizontalAlignment="Center" />
+        </ContentControl>
+    </Grid>
+</Window>
